reorganize spot-related entities and update package structure
