+++ venus/src/model/interface/spot/search-spot/searchSpotDtoPage.ts
@@ -1,4 +1,4 @@
 import SearchSpotDto from "./searchSpotDto";
-import Pageable from "../../pageable";
+import Pageable from "../../../pageable";
 
 export default interface SearchSpotDtoPage extends Pageable<SearchSpotDto> {}
