+++ vulcanus/src/main/java/com/merkury/vulcanus/model/dtos/spot/comment/SpotCommentMediaDto.java
@@ -1,4 +1,4 @@
-package com.merkury.vulcanus.model.dtos.comment;
+package com.merkury.vulcanus.model.dtos.spot.comment;
 
 import com.merkury.vulcanus.model.enums.GenericMediaType;
 import jakarta.validation.constraints.NotBlank;
