+++ vulcanus/src/main/java/com/merkury/vulcanus/model/entities/UserEntity.java
@@ -5,6 +5,9 @@ import com.merkury.vulcanus.model.entities.chat.Chat;
 import com.merkury.vulcanus.model.entities.chat.ChatInvitation;
 import com.merkury.vulcanus.model.entities.chat.ChatMessage;
 import com.merkury.vulcanus.model.entities.chat.ChatParticipant;
+import com.merkury.vulcanus.model.entities.spot.FavoriteSpot;
+import com.merkury.vulcanus.model.entities.spot.SpotComment;
+import com.merkury.vulcanus.model.entities.spot.SpotMedia;
 import com.merkury.vulcanus.model.enums.Provider;
 import com.merkury.vulcanus.model.enums.UserRole;
 import jakarta.persistence.*;
