+++ vulcanus/src/main/java/com/merkury/vulcanus/model/entities/spot/SpotMedia.java
@@ -1,5 +1,6 @@
-package com.merkury.vulcanus.model.entities;
+package com.merkury.vulcanus.model.entities.spot;
 
+import com.merkury.vulcanus.model.entities.UserEntity;
 import com.merkury.vulcanus.model.enums.GenericMediaType;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
