+++ vulcanus/src/main/java/com/merkury/vulcanus/model/mappers/spot/SpotMapper.java
@@ -1,4 +1,4 @@
-package com.merkury.vulcanus.model.mappers;
+package com.merkury.vulcanus.model.mappers.spot;
 
 import com.merkury.vulcanus.model.dtos.spot.FullSpotDto;
 import com.merkury.vulcanus.model.dtos.spot.GeneralSpotDto;
@@ -6,9 +6,9 @@ import com.merkury.vulcanus.model.dtos.spot.SearchSpotDto;
 import com.merkury.vulcanus.model.dtos.spot.SpotDetailsDto;
 import com.merkury.vulcanus.model.embeddable.BorderPoint;
 import com.merkury.vulcanus.model.dtos.spot.coordinates.SpotCoordinatesDto;
-import com.merkury.vulcanus.model.entities.SpotComment;
-import com.merkury.vulcanus.model.entities.Spot;
-import com.merkury.vulcanus.model.entities.SpotMedia;
+import com.merkury.vulcanus.model.entities.spot.SpotComment;
+import com.merkury.vulcanus.model.entities.spot.Spot;
+import com.merkury.vulcanus.model.entities.spot.SpotMedia;
 import com.merkury.vulcanus.model.enums.GenericMediaType;
 import jakarta.validation.constraints.NotNull;
