+++ vulcanus/src/main/java/com/merkury/vulcanus/model/repositories/SpotCommentMediaRepository.java
@@ -1,6 +1,6 @@
 package com.merkury.vulcanus.model.repositories;
 
-import com.merkury.vulcanus.model.entities.SpotCommentMedia;
+import com.merkury.vulcanus.model.entities.spot.SpotCommentMedia;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.query.Param;
